requirements.txt
fastapi
uvicorn[standard]
requests
streamlit
transformers
wikipedia
duckduckgo-search
python-dotenv
python-multipart
PyPDF2
pdfplumber
